extract data request download function
